Migrate ManagedServiceFactory user directory providers to DS factory configuration
Replaces the five ManagedServiceFactory-based user directory provider
factories (Moodle, Sakai, Brightspace, Studip, LDAP) with OSGi
Declarative Services' native factory configuration support
(configurationPolicy = REQUIRE), as part of eliminating managed
services in favor of Declarative Services (opencast/opencast#8004).

Each *UserProviderFactory class is removed; its configuration parsing
and service registration move directly onto the corresponding
*UserProviderInstance class, which becomes the DS component itself.
The runtime now creates, updates, and destroys one component instance
per factory configuration file automatically, replacing the manual
ServiceRegistration bookkeeping the factories used to do by hand.
Existing configuration files (e.g.
org.opencastproject.userdirectory.moodle-default.cfg) continue to
work unchanged, since that naming convention was always driven by
Configuration Admin's factory configuration mechanism, not by
ManagedServiceFactory specifically.

CatalogUIAdapterFactory is intentionally left on ManagedServiceFactory:
it selects between four different adapter classes based on
configuration content, which doesn't map onto DS's one-class-per-PID
factory configuration model without breaking existing custom catalog
configuration files.

This commit was largely AI generated.
