+++ modules/userdirectory-studip/src/main/java/org/opencastproject/userdirectory/studip/StudipUserProviderInstance.java
@@ -26,10 +26,12 @@ import org.opencastproject.security.api.JaxbOrganization;
 import org.opencastproject.security.api.JaxbRole;
 import org.opencastproject.security.api.JaxbUser;
 import org.opencastproject.security.api.Organization;
+import org.opencastproject.security.api.OrganizationDirectoryService;
 import org.opencastproject.security.api.Role;
 import org.opencastproject.security.api.RoleProvider;
 import org.opencastproject.security.api.User;
 import org.opencastproject.security.api.UserProvider;
+import org.opencastproject.util.NotFoundException;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
@@ -37,6 +39,7 @@ import com.google.common.cache.LoadingCache;
 import com.google.common.util.concurrent.ExecutionError;
 import com.google.common.util.concurrent.UncheckedExecutionException;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
@@ -47,6 +50,12 @@ import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
+import org.osgi.service.cm.ConfigurationException;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.component.annotations.Modified;
+import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,6 +70,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
@@ -69,12 +79,36 @@ import java.util.concurrent.atomic.AtomicLong;
 /**
  * A UserProvider that reads user roles from Studip.
  */
+@Component(
+    immediate = true,
+    configurationPid = "org.opencastproject.userdirectory.studip",
+    configurationPolicy = ConfigurationPolicy.REQUIRE,
+    service = { UserProvider.class, RoleProvider.class },
+    property = {
+        "service.description=Provides Studip user directory instances"
+    }
+)
 public class StudipUserProviderInstance implements UserProvider, RoleProvider {
 
   public static final String PROVIDER_NAME = "studip";
 
   private static final String OC_USERAGENT = "Opencast";
   private static final String STUDIP_GROUP = Group.ROLE_PREFIX + "STUDIP";
+
+  /** The key to look up the organization identifer in the service configuration properties */
+  private static final String ORGANIZATION_KEY = "org.opencastproject.userdirectory.studip.org";
+
+  /** The key to look up the URL of the Studip instance */
+  private static final String STUDIP_URL_KEY = "org.opencastproject.userdirectory.studip.url";
+
+  /** The key to look up the user token to use for performing searches. */
+  private static final String STUDIP_TOKEN_KEY = "org.opencastproject.userdirectory.studip.token";
+
+  /** The key to look up the number of user records to cache */
+  private static final String CACHE_SIZE = "org.opencastproject.userdirectory.studip.cache.size";
+
+  /** The key to look up the number of minutes to cache users */
+  private static final String CACHE_EXPIRATION = "org.opencastproject.userdirectory.studip.cache.expiration";
 
   /** The logger */
   private static final Logger logger = LoggerFactory.getLogger(StudipUserProviderInstance.class);
@@ -100,38 +134,75 @@ public class StudipUserProviderInstance implements UserProvider, RoleProvider {
   /** The URL of the Studip instance */
   private String studipToken = null;
 
-  /**
-   * Constructs an Studip user provider with the needed settings.
-   *
-   * @param pid
-   *          the pid of this service
-   * @param organization
-   *          the organization
-   * @param url
-   *          the url of the Studip server
-   * @param token
-   *          the token to authenticate with
-   * @param cacheSize
-   *          the number of users to cache
-   * @param cacheExpiration
-   *          the number of minutes to cache users
-   */
-  public StudipUserProviderInstance(
-      String pid,
-      Organization organization,
-      URI url,
-      String token,
+  /** The organization directory service */
+  private OrganizationDirectoryService orgDirectory;
 
-      int cacheSize,
-      int cacheExpiration
-  ) {
+  /** OSGi DI */
+  @Reference
+  public void setOrgDirectory(OrganizationDirectoryService orgDirectory) {
+    this.orgDirectory = orgDirectory;
+  }
 
-    this.organization = organization;
-    this.studipUrl = url;
-    this.studipToken = token;
+  @Activate
+  @Modified
+  public void updated(Map<String, Object> properties) throws ConfigurationException {
+    String organizationId = (String) properties.get(ORGANIZATION_KEY);
+    if (StringUtils.isBlank(organizationId)) {
+      throw new ConfigurationException(ORGANIZATION_KEY, "is not set");
+    }
 
-    logger.info("Creating new StudipUserProviderInstance(pid={}, url={}, cacheSize={}, cacheExpiration={})",
-                 pid, url, cacheSize, cacheExpiration);
+    URI url;
+    try {
+      url = new URI((String) properties.get(STUDIP_URL_KEY));
+    } catch (URISyntaxException e) {
+      throw new ConfigurationException(STUDIP_URL_KEY, "URI is invalid", e);
+    }
+    if (StringUtils.isBlank(url.toString())) {
+      throw new ConfigurationException(STUDIP_URL_KEY, "is not set");
+    }
+
+    String token = (String) properties.get(STUDIP_TOKEN_KEY);
+    if (StringUtils.isBlank(token)) {
+      throw new ConfigurationException(STUDIP_TOKEN_KEY, "is not set");
+    }
+
+    int cacheSize = 1000;
+    try {
+      if (properties.get(CACHE_SIZE) != null) {
+        Integer configuredCacheSize = Integer.parseInt(properties.get(CACHE_SIZE).toString());
+        if (configuredCacheSize != null) {
+          cacheSize = configuredCacheSize.intValue();
+        }
+      }
+    } catch (Exception e) {
+      throw new ConfigurationException("{} could not be loaded", CACHE_SIZE);
+    }
+
+
+    int cacheExpiration = 60;
+    try {
+      if (properties.get(CACHE_EXPIRATION) != null) {
+        Integer configuredCacheExpiration = Integer.parseInt(properties.get(CACHE_EXPIRATION).toString());
+        if (configuredCacheExpiration != null) {
+          cacheExpiration = configuredCacheExpiration.intValue();
+        }
+      }
+    } catch (Exception e) {
+      throw new ConfigurationException("{} could not be loaded", CACHE_EXPIRATION);
+    }
+
+    try {
+      organization = orgDirectory.getOrganization(organizationId);
+    } catch (NotFoundException e) {
+      logger.warn("Organization {} not found!", organizationId);
+      throw new ConfigurationException(ORGANIZATION_KEY, "not found");
+    }
+
+    studipUrl = url;
+    studipToken = token;
+
+    logger.info("Configured StudipUserProviderInstance(url={}, cacheSize={}, cacheExpiration={})",
+                 url, cacheSize, cacheExpiration);
 
     // Setup the caches
     cache = CacheBuilder.newBuilder().maximumSize(cacheSize).expireAfterWrite(cacheExpiration, TimeUnit.MINUTES)
